requirements.txt
Django==5.2.3
djangorestframework==3.16.0
djoser==2.3.1
django-cors-headers==4.7.0
django-environ==0.12.0
dj-database-url==3.0.0
psycopg2-binary==2.9.10
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
python-decouple==3.8
PyJWT==2.9.0
requests==2.32.4
requests-oauthlib==2.0.0
gunicorn
whitenoise
cloudinary
django-cloudinary-storage
django-storages
pillow
dj-stripe
drf-spectacular
django-filter
cwapi3d
social-auth-app-django
